Rewrite the _shouldAnimate() patching, update readme
Since the CinnamonMagicLamp effect will soon change to patch _shouldAnimate() in the same way that CinnamonBurnMyWindows does, we need a way to coordinate the patching so that the two extensions can coexist. This change will introduce a ShouldAnimateManager class that will be used in both extensions. This class will manage the patching of _shouldAnimate() and allow both extensions to intercept different calling contexts of _shouldAnimate().