Added options to define special rules for handling dialog windows
Added options in the configuration that allows you to define which effects will be used (if any) for dialog windows (i.e. file open dialog windows). This will allow you to use some of the fancy effects for normal windows and the more subtle effects (focus or glide) for all the dialog windows that are opened and closed more frequently.