Add Magic Lamp Effect and an Abbout tab in the config
This is a port of the hermes83 Magic Lamp effect code (the same code used
by CinnamonMagicLamp extension) into a Burn-My-Windows effect. This
allows you to use just one extension to get all effects. Currently there
is an annoying bit of flashing at the start of the minimize effect that
I am working on. You can also use the effect for open/close events, but
the target of the animation might not make sense in those cases. One day
I hope to make it animage from/to the mouse location for open/close
events.