Fix for Mushroom when using a translation, new Magic Lamp settings
1. Fix Mushroom effects presets not working when using a translation

Thanks to GuillaumeV-cemea for spotting the issue with the Fire effect and for fixing it. The same issue he detected applies to the Mushroom effect as well so it needs to be fixed as well. I decided to make the code a bit less error prone by using constants to define the preset types, removing the need to compare against strings which was the cause of the problem in the first place. The down side of this way of fixing the problem is that the existing user settings for Fire and Mushroom preset values will be reset to default after this change is applied. So you will need to change the preset selection back to your preference if you were using something other than the default preset before upgrading to this version.

2. Added "Open/Close Edge" and "Open/Close Offset" options under the Magic Lamp effect settings. These options allow you to set the monitor edge location where the Magic Lamp effect will animate from/to for open and close events. The default will be the "Closest edge to mouse pointer" but you can set it to a specific edge and offset. For example, you could direct the animations to towards the Main Menu location. These options only apply to Open/Close events as the minimize/unminimize events will continue to use the location of the windows icon on the window-list as it's source/target location.

3. Make the custom scale widget show tooltips when hovering over the widgets label rather than just when hovering the scale bar.